Proper build along with better main window
